Fixed reseeding int problem

+++ sph/code_running.py
@@ -18,7 +18,7 @@ comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
 size = comm.Get_size()
 
-np.random.seed(seed=time.time() * rank)
+np.random.seed(seed=int(time.time() * rank))
 
 G = constants.G
 k = constants.Boltzmann
